[refact] Do a big refact of card code

+++ client/src/app/api.ts
@@ -1,12 +1,5 @@
 import axios from "axios"
-import { Dispatch, SetStateAction } from "react"
-import {
-    Claim,
-    CookieBreak,
-    UpdateBreaksFn,
-    UpdateClaimsFn,
-    User,
-} from "./structs"
+import { Claim, CookieBreak, UpdateFn, User } from "./structs"
 import { Data, SetState } from "./page"
 
 const dateOrUndefined = (datetime: string | undefined) =>
@@ -30,7 +23,7 @@ const responseToClaim = (c: any, breaks: CookieBreak[]) => ({
     id: c.id,
     date: new Date(c.claim_date),
     breaks: c.breaks_claimed.map((id: number) =>
-        breaks.find((cb) => cb.id === id)
+        breaks.find((cookieBreak) => cookieBreak.id === id)
     ),
     amount: parseFloat(c.claim_amount),
     reimbursed: dateOrUndefined(c.claim_reimbursed),
@@ -129,59 +122,63 @@ const getHeaders = (user: User | undefined) =>
           }
 
 export const announceBreak = async (
-    user: User,
-    cb: CookieBreak,
-    updateBreaks: UpdateBreaksFn,
+    user: User | undefined,
+    cookieBreak: CookieBreak,
+    updateBreaks: UpdateFn<CookieBreak>,
     setLoadingCard: SetState<boolean>
 ) => {
-    let endpoint = `/api/breaks/announce`
-    let config = {
-        params: {
-            break_id: cb.id,
-        },
-        headers: getHeaders(user),
+    if (user) {
+        let endpoint = `/api/breaks/announce`
+        let config = {
+            params: {
+                break_id: cookieBreak.id,
+            },
+            headers: getHeaders(user),
+        }
+        setLoadingCard(true)
+        let response = await axios.post(endpoint, null, config)
+        let responseData = response.data
+        updateBreaks([responseToBreak(responseData)], [])
+        setTimeout(() => setLoadingCard(false), 1)
     }
-    setLoadingCard(true)
-    let response = await axios.post(endpoint, null, config)
-    let responseData = response.data
-    updateBreaks([responseToBreak(responseData)], [])
-    setTimeout(() => setLoadingCard(false), 1)
 }
 
 export const reimburseBreak = async (
-    user: User,
-    id: number,
+    user: User | undefined,
+    cookieBreak: CookieBreak,
     cost: number,
-    updateBreaks: UpdateBreaksFn,
+    updateBreaks: UpdateFn<CookieBreak>,
     setLoadingCard: SetState<boolean>
 ) => {
-    let endpoint = `/api/breaks/reimburse`
-    let config = {
-        params: {
-            break_id: id,
-            cost,
-        },
-        headers: getHeaders(user),
+    if (user) {
+        let endpoint = `/api/breaks/reimburse`
+        let config = {
+            params: {
+                break_id: cookieBreak.id,
+                cost,
+            },
+            headers: getHeaders(user),
+        }
+        setLoadingCard(true)
+        let response = await axios.post(endpoint, null, config)
+        let responseData = response.data
+        updateBreaks([responseToBreak(responseData)], [])
+        setTimeout(() => setLoadingCard(false), 1)
     }
-    setLoadingCard(true)
-    let response = await axios.post(endpoint, null, config)
-    let responseData = response.data
-    updateBreaks([responseToBreak(responseData)], [])
-    setTimeout(() => setLoadingCard(false), 1)
 }
 
 export const setHost = async (
     user: User,
-    cb: CookieBreak,
+    cookieBreak: CookieBreak,
     host: string,
-    updateBreaks: UpdateBreaksFn,
+    updateBreaks: UpdateFn<CookieBreak>,
     setLoadingCard: SetState<boolean>
 ) => {
     let actualHost = host === "" ? undefined : host
     let endpoint = `/api/breaks/host`
     let config = {
         params: {
-            break_id: cb.id,
+            break_id: cookieBreak.id,
             host_name: actualHost,
         },
         headers: getHeaders(user),
@@ -194,48 +191,52 @@ export const setHost = async (
 }
 
 export const setHoliday = async (
-    user: User,
-    cb: CookieBreak,
+    user: User | undefined,
+    cookieBreak: CookieBreak,
     reason: string | undefined,
-    updateBreaks: UpdateBreaksFn,
+    updateBreaks: UpdateFn<CookieBreak>,
     setLoadingCard: SetState<boolean>
 ) => {
-    let endpoint = `/api/breaks/holiday`
-    let config = {
-        params: {
-            break_id: cb.id,
-            reason,
-        },
-        headers: getHeaders(user),
+    if (user) {
+        let endpoint = `/api/breaks/holiday`
+        let config = {
+            params: {
+                break_id: cookieBreak.id,
+                reason,
+            },
+            headers: getHeaders(user),
+        }
+        setLoadingCard(true)
+        let response = await axios.post(endpoint, null, config)
+        let responseData = response.data
+        updateBreaks([responseToBreak(responseData)], [])
+        setTimeout(() => setLoadingCard(false), 1)
     }
-    setLoadingCard(true)
-    let response = await axios.post(endpoint, null, config)
-    let responseData = response.data
-    updateBreaks([responseToBreak(responseData)], [])
-    setTimeout(() => setLoadingCard(false), 1)
 }
 
 export const deleteBreak = async (
-    user: User,
-    cb: CookieBreak,
-    updateBreaks: UpdateBreaksFn,
+    user: User | undefined,
+    cookieBreak: CookieBreak,
+    updateBreaks: UpdateFn<CookieBreak>,
     setLoadingCard: SetState<boolean>
 ) => {
-    let endpoint = `/api/breaks/${cb.id}`
-    let config = {
-        headers: getHeaders(user),
+    if (user) {
+        let endpoint = `/api/breaks/${cookieBreak.id}`
+        let config = {
+            headers: getHeaders(user),
+        }
+        setLoadingCard(true)
+        await axios.delete(endpoint, config)
+        updateBreaks([], [cookieBreak])
+        setTimeout(() => setLoadingCard(false), 1)
     }
-    setLoadingCard(true)
-    await axios.delete(endpoint, config)
-    updateBreaks([], [cb])
-    setTimeout(() => setLoadingCard(false), 1)
 }
 
 export const submitClaim = async (
     user: User,
     cbs: CookieBreak[],
-    updateBreaks: UpdateBreaksFn,
-    updateClaims: UpdateClaimsFn,
+    updateBreaks: UpdateFn<CookieBreak>,
+    updateClaims: UpdateFn<Claim>,
     setLoadingCards: (loading: boolean) => void
 ) => {
     let endpoint = `api/claims/claim`
@@ -245,7 +246,7 @@ export const submitClaim = async (
     setLoadingCards(true)
     let response = await axios.post(
         endpoint,
-        cbs.map((cb) => cb.id),
+        cbs.map((cookieBreak) => cookieBreak.id),
         config
     )
     let responseData = response.data
@@ -260,7 +261,7 @@ export const completeClaim = async (
     user: User,
     claim: Claim,
     breaks: CookieBreak[],
-    updateClaims: UpdateClaimsFn,
+    updateClaims: UpdateFn<Claim>,
     setLoadingCard: SetState<boolean>
 ) => {
     let endpoint = `api/claims/success`
